Add detection for selecting the same slot twice
A new field, lastSlot, has been added to track the last selected slot in InvalidPacketDetection.java. The function has been updated to detect and respond to cases where the same slot is selected twice in sequence, a behaviour that is considered invalid. The player is kicked upon detection of this instance.