Add a utility for checking latest plugin version
Introduced a new utility class, `UpdateChecker`, to fetch the latest plugin version from GitHub. This is used in `Sierra` and `DataManager` to compare current plugin version against the latest release, and warn if new updates are available. This feature aims to help keep the plugin up-to-date, as any outdated versions will now prompt notifications.