Add monitoring, timing, and ping processing features
This commit introduces several new features related to performance monitoring and timing. Some include the addition of PingProcessor and TimingProcessor to handle ping and timing processing for the players. Also, monitor information related to the player is now printed using the MonitorCommand implementation. Furthermore, the SierraTiming class provides methods to measure and track the timing of events while the Timing interface represents this in a general way. Finally, an alert history on violations was also added, which shows ping information.