Simplify InvalidPacketDetection checks and logging
Removed a redundancy check that compared the entity id of a user to the entity id caught in the event, as it was unnecessary. Simplified the debug information logging to only display the jump boost details. This refactor aims to streamline the invalid packet detection and make the debug information logging more straightforward.