Refactor server version check method name
The method responsible for checking if the server version is supported or not has been renamed from 'isSupportedVersion' to 'isSupportedServerVersion' for better clarity. This name change helps to differentiate it from checks on the client version. As a result of its name change, all its usages across the file have been updated accordingly.