Add username validation to packet handling
A new boolean field 'nameChecked' has been added to the PlayerData class to keep track of whether a username has been validated. In the PacketListener class, a method 'checkHandling' has been implemented to check and validate the username of a user when a packet is received. If the username is invalid, the event is canceled and the player is kicked. Furthermore, a method 'isValidUsername' has been introduced to determine the validity of a username based on a specific pattern.