Update version and fixed command blocking method
The build.gradle and plugin.yml files have been updated with a new minor version. The BlockedCommand.java was also modified to improve the command blocking method. Instead of simply cancelling the event when a match with the MVC_PATTERN is found, a violation is recorded with the matching string logged in the debug information.