Update default permissions, gradle dependencies and modify violation responses
This commit updates the default permissions for the 'sierra.command.alerts' in plugin.yml and changes the PacketEvents dependency version in the build.gradle. It also introduces a mitigation response for certain violations in 'InvalidPacketDetection.java', replacing the previous kick punishment. The item causing the violation will be removed if the punishment configuration is set to HARD. Additionally, it includes a minor enhancement in code formatting in the 'FormatUtils.java'.