Add alert management functionality
A new functionality has been added to manage alert settings and mitigations for each player. The system now allows toggling the enabled state of an alert or a mitigation using implemented commands. The AlertSettings interface and its AbstractAlertSetting implementation provide basic manipulations with alerts, such as checking if an alert is enabled and changing its state. The SierraUser method was updated to retrieve AlertSettings instead of a boolean indicating whether the user has any alerts or not. Appropriate modifications were made in related classes.