Update Sierra plugin to version 1.2.5 and enhance user messaging
The Sierra plugin was updated to version 1.2.5, featuring adjustments to SierraDataManager.java, build.gradle, and the plugin.yml file. The User Messaging was improved in SierraDataManager.java, particularly, changing the recipient of the message from a player to user and enhancing the messaging content. Also, a todo comment was added in TimerDetection.java for a potential check in the next release.