Add metrics for kicks and bans
Added two new metrics to the Sierra application to track the number of kicks and bans. Initialized these metrics in the SierraDataManager class and updated the kick and ban methods in the PlayerData class to increment the respective counters. These metrics are represented in a custom chart by the PacketEvents API.