Add NBT length check
In this update, a new function `checkGenericNBTLimit` was introduced that checks the length of the NBT data in the given ItemStack against a defined limit and takes appropriate actions if the limit is exceeded. This limit check has been incorporated in multiple places within the code. The function will help eliminate many cheap crashers and improve overall data security. A corresponding configuration `generic-nbt-limit` has been added in `sierra.yml`.