Updated InfoCommand to use ConfigValue for messages
The InfoCommand class has been updated to use the ConfigValue class for message construction. This allows more flexibility for customization, as admins can modify the phrases directly through a configuration file. Also, this makes it easier to support multiple languages in the future. The changes also include the part of pulling the message out of a yml file.