Refactor messages and update color config
The command messages in CommandHelper, SierraCommand, and InfoCommand have been refactored to include a customizable offset-color setting. This allows for more flexibility in setting the color scheme for output messages. A new field 'offset-color' has been added to the config file to facilitate this change. These changes provide greater customization for users in how they can display the plugin's messages.