Update SierraLoader to version 1.0.1 and enhance error handling
This commit bumps the version of SierraLoader to 1.0.1 and improves the error handling in the load function. The writeInputStreamFromURL function is deleted and its IO process is directly included in the onEnable function, where more proper log messages can help users understand the current progress during plugin loading. Also, a syntax modification that added an empty line was included in InvalidPacketDetection.java.