Add plugin exploits check and refine violation responses
Updated the BlockedCommand class to include a method that checks for plugin exploits via given commands. Made code optimizations in both the BlockedCommand and TimerDetection classes for violations, where the punishment type is now directly dependent on the number of violations. Minor changes have also been made in the InvalidPacketDetection class and sierra.yml.