Add violation count and kick functionality for check types
Added a new functionality to track and increment the violation count for specific check types. A new HashMap variable "violationCount" was introduced in the SierraDataManager class to store this data. In addition, revisions were made in the SierraDetection and InvalidMoveDetection classes to make use of this new functionality and also handle entity locations more effectively.