Refactor punishment system and reduce code duplication
This commit refactors the punishment system in InvalidPacketDetection, and reduces code duplication in Ticker and InvalidMoveDetection. The punishment determination in InvalidPacketDetection is now inline, basing on the amount of violations. In Ticker, the check for LATENCY_ABUSE is removed, since this check is not specific and is executed for every check. The InvalidMoveDetection code is more concise with the checkDelta method call placed in the same line as the condition.