Add IP blocking feature to manage invalid connections
A new feature has been introduced to block IP addresses if they have been associated with bad behavior. The `AddressStorage` class has been created to store and manage IP addresses along with their added times. An IP address is considered invalid and gets blocked if it was recently punished, this logic is added in `SierraDetection` and `SierraDataManager`. The "block-connections-after-ban" property is introduced in the Sierra configuration to control this feature. The blocked address will stay inactive for a specified time set in "connection-block-time".