Add active check types chart and improve kick handling
Refactored the metrics initializer in Sierra.java to include an AdvancedPie chart for active check types. Also moved the addKick call in SierraDetection.java to only occur if the violationDocument's punishType is not MITIGATE. Furthermore, the addKick method in SierraDataManager.java now eliminates spaces in checkType names when storing them in the violationCount map.