Add Sierra application initialization logic
This commit introduces various classes responsible for the initialization and cleanup of the Sierra application. This includes enabling and disabling of PacketEvents API, bStats metrics setup, starting and finishing application initialization processes and implementing various listeners for packet receive events. The initialization procedure includes the usage of the InitManager class to manage loading, starting, and stopping the initialization process. In addition, the PacketReceiveListener class is introduced to help manage the handling of received packets. Changes also include the deletion of the PacketListener class.