Refactor a significant portion of the Sierra plugin
This commit includes changes to classes and functions across different areas of the plugin. The most notable change is the transition from using `ClassToInstanceMap` to `ArrayList` for managing initialization and packet checks, with corresponding modifications in InitManager and CheckManager classes. Additionally, tweaks were made to the ordering of function calls, descriptions were updated in plugin.yml, and irrelevant imports were removed. The build.gradle file was also adjusted, and several bug fixes and corrections were applied to a variety of java files.