Update event handling and data management methods
Refactored the event handling from `onUserLogin` to `onUserConnect` and updated the `addPlayerData` method to no longer need a player object. Made several changes to game mode handling, including moving this responsibility to the `Ticker` class. Also removed a redundant game mode check from the `ProtocolValidation` class. This simplifies data management and ensures that the game mode is always up-to-date.