Add mitigation history and improve packet validation
Added a method to create a mitigation history entry for a user. This registers histories of mitigations applied to users. Improved packet validation by adding null checks and introduced a method to check block placement distance to validate user actions better. Extended type penalization with the MITIGATE option and enhanced History enum with the Mitigate option.