Update Maven dependencies and improve permission handling
This commit changes the Maven repository and version for PacketEvents. It also adjusts the default permission for 'sierra.bypass' and refactors the permission handling logic. Now the user's bypass permission is checked, instead of the 'op' status. Unnecessary code related to block placement distance check in 'ProtocolValidation' is also removed.