Add client version tracking to history logs
The client version is being tracked in the HistoryDocument and reflected in the history logs for both punishment and mitigation events. This update also includes changes in SierraDataManager, PacketReceiveListener, History API, SierraDetection, and HistoryCommand to accommodate this new value. Changes in the history entry message structures have also been reflected in the messages.yml file.