Add InfiniteParkour compatibility with model check disabling
New file InfiniteParkourDescriptor.java was created to handle compatibility with the InfiniteParkour plugin. This includes checking for common issues and disabling the model check if identified, due to faulty NBT data in the plugin. Adjustments were also made in CompatibilityHandler.java to include InfiniteParkourDescriptor. SierraDataManager.java now holds a flag to control skipping the model check, which is used in ProtocolValidation.java.