Added transaction based timer check
This commit involves extensive changes to the MovementValidation class, including a significant modification of the handleLatencyAbuse function to enhance the efficiency of timer cheats prevention. In assistance to this, the calculateResult function was added to the FormatUtils.java class. Other changes include refactoring and removal of redundant variables and functions, aimed at improving the code readability and performance.