Refactor HistoryDocument and HistoryCommand classes
The code in the HistoryDocument and HistoryCommand classes has been cleaned up for simplicity and readability. Superfluous comments in the HistoryDocument class were removed and the 'Data' annotation was added as a shorthand for generating standard getters, setters, equals, etc. In the HistoryCommand class, the sequence of commands was restructured to enhance debuggability and code maintenance.