Add user locale information to Sierra plugin
This update includes a new feature to the Sierra plugin where it collects the user's locale information. The changes affect multiple files, notable modifications involve adding a new 'locale' field to the SierraUser class and incorporating this locale information into the informational messages displayed to the user. The version of the Sierra plugin has also been updated to reflect these changes.