Add configuration to skip negative experience check
A new configuration option `skip-negative-experience-check` has been added to the sierra.yml file. This is used in the ProtocolValidation class to optionally bypass the check for negative experience values in the `checkSetExperience` method. A minor adjustment has also been made to the `handleAnvilInventory` method to remove its need for the `playerData` parameter.