Refactor code for history document creation and event publishing
The code in the SierraDataManager and SierraDetection classes has been refactored for easier readability and maintenance. In SierraDataManager, the creation of HistoryDocument and the firing of AsyncHistoryCreateEvent have been moved inside the runNow method. This change also includes conditionally adding the document to histories only if the event is not cancelled. In SierraDetection, numerous lines have been combined and reformatted for clarity and succinctness.